backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.7.0
agno==2.0.5
google-generativeai==0.8.3
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
anyio==4.8.0
python-dateutil==2.9.0.post0
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0